// projects/ui/src/lib/components/po-page/po-page-detail/po-page-detail-literals.interface.ts
/**
 * @usedBy PoPageDetailComponent
 *
 * @description
 *
 * Interface para definição das literais usadas no `po-page-detail`.
 */
export interface PoPageDetailLiterals {
  /** Label da ação `back`. */
  back?: string;

  /** Label da ação `edit`. */
  edit?: string;

  /** Label da ação `remove`. */
  remove?: string;
}